refactor: extract chord step encoder controls
